// src/types.ts
export type TabType = 
  | 'beranda'
  | 'semua-turnamen'
  | 'ff'
  | 'mlbb'
  | 'total-hadiah'
  | 'pengumuman'
  | 'riwayat'
  | 'status-pembayaran'
  | 'form-pendaftaran'
  | 'pengaturan-umum'
  | 'bantuan'
  | 'arsip'
  | 'bagikan'
  | 'cara-daftar'
  | 'aturan'
  | 'tim'
  | 'info-match'
  | 'grup'
  | 'laporan'
  | 'kontak'
  | 'login'
  | 'prediksi'
  | 'saldo'
  | 'admin'
  | 'topup-game'
  | 'unduh-apk'
  | 'donasi'
  | 'gemini-ai';

export interface DonationRecord {
  id: string;
  donorName: string;
  isAnonymous: boolean;
  amount: number;
  message?: string;
  paymentMethod: 'Saweria QRIS';
  status: 'BERHASIL' | 'GAGAL' | 'MENUNGGU_VERIFIKASI';
  createdAt: string;
  timestamp: number;
}

export type UserRole = 'guest' | 'peserta' | 'admin';
// ...
export interface DeviceInfo {
  deviceId: string;
  deviceType: 'Mobile' | 'Tablet' | 'Desktop';
  os: string;
  browser: string;
  screenRes: string;
  userAgent: string;
  lastLogin: string;
  isNotificationEnabled: boolean;
  notificationPermission: 'granted' | 'denied' | 'default';
}

export interface UserAccount {
  id?: string;
  name: string;
  email: string;
  password?: string;
  phone?: string;
  role: UserRole;
  avatarUrl?: string;
  googleId?: string;
  isSuperAdmin?: boolean;
  status?: 'Active' | 'Blacklisted';
  teamName?: string;
  registeredAt?: string;
  deviceInfo?: DeviceInfo;
  failedAttempts?: number;
  isLocked?: boolean;
  lockUntil?: string;
  resetCode?: string;
  userKey?: string;
}